Reformatted the output
Beatrice report some bug, need to check

+++ inc/storage.hpp
@@ -58,7 +58,7 @@
 	// threshold, r2,  num_snps, p, coefficient r2 adjust, number of better
 	typedef std::tuple<double, double, size_t, double, double, size_t, double> PRSice_result;
 	//  threshold, r2, num_snps, coefficient, pvalue (The r2 is for determining if it is the best), number of better
-	typedef std::tuple<double, double, size_t, double, double, size_t> PRSice_best;
+	typedef std::tuple<double, double, size_t, double, double, int> PRSice_best;
     typedef std::tuple<std::string, char, std::string, std::string> help;
     typedef std::tuple<std::string, size_t, size_t> boundary;
     typedef std::tuple<std::string, int, size_t> snp_link;
